Refactoring remarks pull request
