+++ my-project/src/app/index.js
@@ -1,16 +1,18 @@
 import React, { Component } from 'react';
-import LoginForm from '../components/LoginForm/index'
+import LoginForm from '../components/LoginForm/LoginForm'
+import submit from '../components/LoginForm/submit';
 import './style.scss';
 
 class App extends Component {
-    submit = values => {
-        alert(JSON.stringify(values));
-    };
+    // submit = values => {
+    //     console.log(JSON.stringify(values));
+    //     alert(JSON.stringify(values));
+    // };
     render() {
         return (
             <>
                 <h1 className="title">Sing in</h1>
-                <LoginForm onSubmit={this.submit}></LoginForm>
+                <LoginForm onSubmit={submit}/>
             </>
         )
     }
