--- my-project/src/components/Field/index.js
@@ -1,19 +1,0 @@
-import React from 'react';
-import './style.scss';
-
-const myInput = (props) => {
-    const { input, type, placeholder, meta } = props;
-    return (
-        <>
-            <input {...props.input} type={props.type} placeholder={props.placeholder} />
-            {meta.error &&
-                meta.touched &&
-                <div>
-                    {meta.error}
-                </div>
-            }
-        </>
-    )
-}
-
-export default myInput;
