+++ my-project/src/components/LoginForm/LoginForm.scss
@@ -1,0 +1,21 @@
+.image {
+  min-height: 540px;
+  position: relative;
+  // background-color: red;
+  background-image: url(image.svg);
+  background-repeat: no-repeat;
+  background-position: 25% 50%;
+}
+.formLogin {
+  position: absolute;
+  right: 10%;
+  top: 30%;
+}
+
+button {
+  margin: 10px 10px 0 0;
+  padding: 10px;
+  background-color: blue;
+  color: white;
+  border: none;
+}
