+++ my-project/src/components/LoginForm/submit.js
@@ -1,0 +1,11 @@
+// function submit(values) {
+//   console.log(JSON.stringify(values));
+//   alert(JSON.stringify(values));
+// }
+
+const submit = values => {
+  console.log(JSON.stringify(values));
+  alert(JSON.stringify(values));
+};
+
+export default submit;
